0.4.3 - Refactoring. Add buttons processing
