+++ src/main/java/halot/nikitazolin/bot/discord/CommandRegistrationService.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import halot.nikitazolin.bot.discord.command.manager.CommandCollector;
+import halot.nikitazolin.bot.discord.command.CommandCollector;
 import halot.nikitazolin.bot.discord.command.model.BotCommand;
 import halot.nikitazolin.bot.discord.jda.JdaMaker;
 import lombok.Getter;
