+++ src/main/java/halot/nikitazolin/bot/discord/listener/JdaListenerService.java
@@ -5,7 +5,6 @@ import java.util.Optional;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
-import halot.nikitazolin.bot.discord.command.manager.CommandEventHandler;
 import halot.nikitazolin.bot.discord.jda.JdaMaker;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,7 +17,7 @@ import net.dv8tion.jda.api.JDA;
 public class JdaListenerService {
 
   private final JdaMaker jdaMaker;
-  private final CommandEventHandler commandEventHandler;
+  private final EventHandler eventHandler;
 
   public void addListeners() {
     Optional<JDA> jda = jdaMaker.getJda();
@@ -28,10 +27,8 @@ public class JdaListenerService {
 
   private void registerListeners(Optional<JDA> jda) {
     jda.ifPresentOrElse(jdaL -> {
-      jdaL.addEventListener(commandEventHandler);
-      log.info("register JDA Listener: " + commandEventHandler);
-//      jdaL.addEventListener(another listener);
+      jdaL.addEventListener(eventHandler);
+      log.info("register event listener: " + eventHandler);
     }, () -> System.out.println("JDA is not present!"));
-
   }
 }
