+++ src/main/java/halot/nikitazolin/bot/init/settings/manager/SettingsSaver.java
@@ -1,4 +1,4 @@
-package halot.nikitazolin.bot.init.authorization.manager;
+package halot.nikitazolin.bot.init.settings.manager;
 
 import java.io.FileWriter;
 import java.io.IOException;
@@ -8,37 +8,37 @@ import org.springframework.stereotype.Component;
 import org.yaml.snakeyaml.DumperOptions;
 import org.yaml.snakeyaml.Yaml;
 
-import halot.nikitazolin.bot.init.authorization.model.AuthorizationData;
+import halot.nikitazolin.bot.init.settings.model.Settings;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 @Component
 @Slf4j
 @RequiredArgsConstructor
-public class AuthorizationSaver {
+public class SettingsSaver {
 
-  private final AuthorizationData authorizationData;
+  private final Settings settings;
 
   public void saveToFile(String filePath) {
-    saveConfig(authorizationData, filePath);
+    saveConfig(settings, filePath);
   }
 
-  private void saveConfig(AuthorizationData authorizationData, String filePath) {
-    log.info("Update config with path: " + filePath);
+  private void saveConfig(Settings settings, String filePath) {
+    log.info("Update settings with path: " + filePath);
     DumperOptions options = new DumperOptions();
     options.setDefaultFlowStyle(DumperOptions.FlowStyle.BLOCK);
     options.setPrettyFlow(true);
     Yaml yaml = new Yaml(options);
 
     try (StringWriter stringWriter = new StringWriter()) {
-      yaml.dump(authorizationData, stringWriter);
+      yaml.dump(settings, stringWriter);
       String output = stringWriter.toString().replaceAll("^!!.*\n", "");
 
       try (FileWriter writer = new FileWriter(filePath)) {
         writer.write(output);
       }
     } catch (IOException e) {
-      log.error("Error writing the secrets file: {}", e);
+      log.error("Error writing the settings file: {}", e);
     }
   }
 }
